Server/requirements.txt
google-cloud-aiplatform
llama-index
google-generativeai
transformers
torch
torchvision
torchaudio
fastapi
sqlalchemy[asyncio]
pgvector
psycopg2-binary
asyncpg
uvicorn[standard]